clean and address comments
